fix: break the no-baseline report import cycle introduced by the file split
`ai-readiness` failed on CI with:
  report.no_baseline -> report.no_baseline_render -> report.no_baseline

Self-inflicted: splitting the SARIF/JUnit renderers out to keep
report/no_baseline.py under the 800-line cap left the renderer importing
NoBaselineDocument back from the compute half. The AI-readiness scan counts
TYPE_CHECKING imports too, so that annotation is a real edge.

It reached CI because after that split I re-ran check_architecture but not
check_ai_readiness -- only the latter catches this.

Fixed per AGENTS.md's own remedy (move the shared logic to a leaf module,
never extend IMPORT_CYCLE_ALLOWLIST): the contract -- NoBaselineDocument,
the schema version, and the supported/unsupported format sets -- now lives
in report/no_baseline_document.py, which both halves depend on and neither
depends back through. That is this package's own document.py-versus-render_*
shape. `report.no_baseline` re-exports all four names, so its documented
import path is unchanged.

Verified: ai-readiness reports no cycle; ruff, mypy (731 files) and
architecture clean; 266 targeted tests pass; all five formats still render
and the compatibility import path resolves.
